Feature google authentication (#21)
* Added Auth Next JS package and fixed tailwind content not working on src/components

* Successully added Google sign in next step is to style the page

* Successfully added route protection using nextauth:

* Added Sign Out button for testing purposes

* Added Callback in Auth For Checking new Users and Adding to the Database

* Updated schema added to attribute to recognize destination

* Added Session on Api route auth and customer

* Added Session on Api route product

* Added Session on Api route product #2

* Revert "Added Session on Api route product #2"

This reverts commit e1c8131039f413fa7ad12b0b56d234ff4be838e7.

* Fixed My Store not Updating when switching from another store

* Reverted id in edit and product due to it already protected by server side props

* Removed index.tsx on store and edit

* Succesfullt redirect page after sign in

* Login Page Done

* Added Head

+++ pages/api/product/deleteUserCart.ts
@@ -1,19 +1,27 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { getSession } from "next-auth/react";
 
 import { prisma } from "../../../lib/prisma";
 
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
   const { id } = req.body;
-  // console.log(id);
+  const session = await getSession({ req });
 
-  try {
-    await prisma.cart.delete({
-      where: {
-        cartID: id,
-      },
-    });
-    res.status(200).end();
-  } catch (e) {
-    console.log(e);
+  if (session) {
+    try {
+      await prisma.cart.delete({
+        where: {
+          cartID: id,
+        },
+      });
+      res.status(200).end();
+    } catch (e) {
+      console.log(e);
+    }
+  } else {
+    return res.status(401);
   }
 }
